refactor: decouple productivity UI from business logic and state
Extracted business logic, data formatting, and side-effects from ProductivityScreen into useProductivityViewModel hook. Split nested view rendering by creating ProductivityHeader and ProductivityEmptyState components.
